feat(mirta-rollup): extract env-loader to package
