+++ packages/mirta-testing/package.json
@@ -41,11 +41,9 @@
     "build:mono": "rollup -c"
   },
   "dependencies": {
+    "@mirta/env-loader": "workspace:*",
     "@mirta/workspace": "workspace:*",
     "mirta": "workspace:*",
     "vitest-mock-extended": "^3.1.0"
-  },
-  "devDependencies": {
-    "@mirta/rollup": "workspace:*"
   }
 }
